Extract file generation from functional tests
